extracting comparison components into their own files
