+++ src/app/providers.tsx
@@ -1,6 +1,6 @@
 "use client";
 
-import CompareModal from "@/components/ComparisonModal/CompareModal";
+import CompareModal from "@/components/ComparisonComponents/ComparisonModal/CompareModal";
 import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
 import { ReactQueryDevtools } from "@tanstack/react-query-devtools";
 import { useState } from "react";
