+++ src/lib/formatting/formatMoney.ts
